+++ android/app/src/main/java/com/brekeke/phonedev/lpc/LpcUtils.java
@@ -2,20 +2,15 @@ package com.brekeke.phonedev.lpc;
 
 import android.annotation.SuppressLint;
 import android.app.ActivityManager;
-import android.app.Notification;
-import android.app.NotificationManager;
-import android.app.PendingIntent;
 import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
 import android.content.ServiceConnection;
 import android.os.IBinder;
-import android.provider.Settings;
 import android.text.TextUtils;
 import android.util.Base64;
 import android.util.Log;
 import com.brekeke.phonedev.R;
-import com.brekeke.phonedev.utils.L;
 import com.facebook.react.ReactApplication;
 import com.facebook.react.ReactInstanceManager;
 import com.facebook.react.bridge.ReadableArray;
@@ -77,34 +72,6 @@ public class LpcUtils {
     ActivityManager.RunningAppProcessInfo myProcess = new ActivityManager.RunningAppProcessInfo();
     ActivityManager.getMyMemoryState(myProcess);
     return myProcess.importance != ActivityManager.RunningAppProcessInfo.IMPORTANCE_FOREGROUND;
-  }
-
-  public static void showIncomingCallNotification(Context appCtx, Intent i) {
-    if (!checkAppInBackground()) {
-      return;
-    }
-    i.setAction(Settings.ACTION_MANAGE_OVERLAY_PERMISSION);
-    // fix: the application crashes when it is in the background state and there is an
-    // incoming call
-    int flags = PendingIntent.FLAG_UPDATE_CURRENT;
-    if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.S) {
-      // |= is used to add a flag without losing other flags already present in the flags
-      // variable.
-      flags |= PendingIntent.FLAG_IMMUTABLE;
-    }
-    PendingIntent pi = PendingIntent.getActivity(appCtx, 0, i, flags);
-
-    Notification notification =
-        new Notification.Builder(appCtx, NOTI_CHANNEL_ID)
-            // fix: the app will crash: "Invalid notification (no valid small icon)"
-            .setSmallIcon(R.mipmap.ic_launcher)
-            .setContentTitle(L.incomingCall())
-            .setCategory(Notification.CATEGORY_CALL)
-            .setFullScreenIntent(pi, true)
-            .build();
-    NotificationManager nm =
-        (NotificationManager) appCtx.getSystemService(Context.NOTIFICATION_SERVICE);
-    nm.notify(TAG, NOTI_ID, notification);
   }
 
   public static final ServiceConnection connection =
@@ -265,17 +232,6 @@ public class LpcUtils {
     sslContext.init(null, trustAllCerts, null);
 
     return sslContext;
-  }
-
-  // TODO: match ssid in background service where it initiates the connection
-  public static Boolean matchSsid(ReadableArray remoteSsid, String localSsid) {
-    for (int i = 0; i < remoteSsid.size(); i++) {
-      ReadableType type = remoteSsid.getType(i);
-      if (type == ReadableType.String && remoteSsid.getString(i).equals(localSsid)) {
-        return true;
-      }
-    }
-    return false;
   }
 
   // Convert ReadableArray to String[]