+++ ios/BrekekePhone/Bridging-Header.h
@@ -16,5 +16,6 @@
 #import "RNCallKeep.h"
 #import "RNSplashScreen.h"
 #import "RNVoipPushNotificationManager.h"
-
+#import <CommonCrypto/CommonCrypto.h>
+#import <RNCAsyncStorage.h>
 #endif