+++ ios/BrekekePhone/Bridging-Header.m
@@ -29,6 +29,18 @@ RCT_EXTERN_METHOD(setProximityMonitoring : (BOOL)enabled)
 RCT_EXTERN_METHOD(systemUptimeMs
                   : (RCTPromiseResolveBlock)resolve rejecter
                   : (RCTPromiseRejectBlock)reject)
+RCT_EXTERN_METHOD(getRingtoneOptions
+                  : (RCTPromiseResolveBlock)resolve rejecter
+                  : (RCTPromiseRejectBlock)reject)
+RCT_EXTERN_METHOD(validateRingtone
+                  : (NSString *)ringtone username
+                  : (NSString *)username tenant
+                  : (NSString *)tenant host
+                  : (NSString *)host port
+                  : (NSString *)port resolver
+                  : (RCTPromiseResolveBlock)resolve rejecter
+                  : (RCTPromiseRejectBlock)reject)
+
 @end
 
 @interface RCT_EXTERN_MODULE (BrekekeEmitter, RCTEventEmitter)