+++ src/components/AccountCreateForm.tsx
@@ -13,6 +13,12 @@ import { intl, intlDebug } from '#/stores/intl'
 import { RnAlert } from '#/stores/RnAlert'
 import type { RingtoneOption } from '#/utils/getRingtoneOptions'
 import { getRingtoneOptions } from '#/utils/getRingtoneOptions'
+import PreviewRingtone from '#/utils/PreviewRingtone'
+import {
+  handleUploadRingtone,
+  saveRingtoneSelection,
+  validateRingtone,
+} from '#/utils/ringtonePicker'
 import { useForm } from '#/utils/useForm'
 import { useStore } from '#/utils/useStore'
 
@@ -31,6 +37,7 @@ export const AccountCreateForm: FC<{
       },
       addingPark: { name: '', number: '' },
       ringtoneOptions: [] as RingtoneOption[],
+      preview: '',
     },
     resetAllFields: () => {
       RnAlert.prompt({
@@ -114,6 +121,16 @@ export const AccountCreateForm: FC<{
     onValidSubmit: () => {
       props.onSave($.account, $.hasUnsavedChanges())
     },
+    onUploadRingtone: async () => {
+      try {
+        await handleUploadRingtone($.ringtoneOptions, options => {
+          $.ringtoneOptions = options
+        })
+      } catch (err) {
+        console.error('AccountCreateForm onUploadRingtone:', err)
+      }
+    },
+    stopPreview: () => ($.preview = ''),
   })
 
   type M0 = ReturnType<typeof m>
@@ -130,6 +147,25 @@ export const AccountCreateForm: FC<{
       })
     }
   }, [$, props.footerLogout])
+
+  const getDropDown = () => {
+    let d = [
+      {
+        label: intl`Reset form`,
+        onPress: $.resetAllFields,
+      },
+    ]
+    if (props.updating && !isWeb) {
+      d = [
+        {
+          label: intl`Select local mp3 as ringtone`,
+          onPress: $.onUploadRingtone,
+        },
+        ...d,
+      ]
+    }
+    return d
+  }
 
   return (
     <Layout
@@ -163,12 +199,7 @@ export const AccountCreateForm: FC<{
                 danger: true,
               },
             ]
-          : [
-              {
-                label: intl`Reset form`,
-                onPress: $.resetAllFields,
-              },
-            ]
+          : getDropDown()
       }
       fabOnBack={props.footerLogout ? undefined : $.onBackBtnPress}
       fabOnNext={props.footerLogout ? undefined : (submitForm as () => void)}
@@ -292,19 +323,29 @@ export const AccountCreateForm: FC<{
             isGroup: true,
             label: intl`Ringtone`,
             hasMargin: true,
-            hidden: props.footerLogout,
+            hidden: isWeb || props.footerLogout,
           },
           {
             disabled: props.footerLogout,
             type: 'RnPicker',
             name: 'ringtone',
             options: $.ringtoneOptions,
-            hidden: props.footerLogout,
+            hidden: isWeb || props.footerLogout,
+            onConfirm: async (value: string) =>
+              await saveRingtoneSelection(value, undefined, $.account),
+            onValueChange: async (value: string) => {
+              const r = await validateRingtone(value, $.account)
+              $.preview = r
+            },
+            onDismiss: $.stopPreview,
           },
         ]}
         k='account'
         onValidSubmit={$.onValidSubmit}
       />
+      {!isWeb && $.preview && (
+        <PreviewRingtone source={$.preview} onFinished={$.stopPreview} />
+      )}
     </Layout>
   )
 })