+++ src/pages/PageSettingsCurrentAccount.tsx
@@ -18,7 +18,7 @@ export const PageSettingsCurrentAccount = observer(() => (
       }
       RnAlert.prompt({
         title: intl`Save Account`,
-        message: intl`Do you want to update your account?\nYou'll need to sign out and then sign in again.`,
+        message: intl`Do you want to update your account?\nYou'll need to sign out and then sign in again`,
         onConfirm: () => {
           ctx.account.upsertAccount(p)
           ctx.nav.goToPageAccountSignIn()