+++ src/stores/accountStore.ts
@@ -107,6 +107,8 @@ export class AccountStore {
     pbxRingtone: defaultRingtone,
   })
 
+  @observable ringtonePicker: RingtonePickerType = {}
+
   loadAccountsFromLocalStorage = async () => {
     const arr = await RnAsyncStorage.getItem('_api_profiles')
     let d: TAccountDataInStorage | undefined
@@ -120,6 +122,7 @@ export class AccountStore {
     }
     if (d) {
       let { profileData: accountData, profiles: accounts } = d
+      const { ringtonePicker } = d
       if (Array.isArray(d)) {
         // lower version compatible
         accounts = d
@@ -138,6 +141,7 @@ export class AccountStore {
           )
         }
         this.accountData = uniqBy(accountData, 'id')
+        this.ringtonePicker = ringtonePicker ?? {}
       })
     }
     resolveFn?.()
@@ -156,6 +160,7 @@ export class AccountStore {
         jsonSafe({
           profiles,
           profileData: this.accountData,
+          ringtonePicker: this.ringtonePicker,
         }),
       )
     } catch (err) {
@@ -178,6 +183,9 @@ export class AccountStore {
     })
     return this._saveAccountsToLocalStorageDebounced()
   }
+
+  saveAccountsToLocalStorageWithoutDebounced = async () =>
+    await this.saveAccountsToLocalStorage()
 
   @action upsertAccount = async (p: Partial<Account>) => {
     const a = this.accounts.find(_ => _.id === p.id)
@@ -291,8 +299,12 @@ export class AccountStore {
     if (d) {
       return d
     }
+    const uniqueId = getAccountUniqueId(a)
+    if (!uniqueId) {
+      throw new Error('Account unique id is undefined')
+    }
     const newD = {
-      id: getAccountUniqueId(a),
+      id: uniqueId,
       accessToken: '',
       recentCalls: [],
       recentChats: [],
@@ -355,6 +367,7 @@ export type RecentCall = AccountData['recentCalls'][0]
 type TAccountDataInStorage = {
   profiles: Account[]
   profileData: AccountData[]
+  ringtonePicker: RingtonePickerType
 }
 
 type LastSignedInId = {
@@ -364,6 +377,10 @@ type LastSignedInId = {
   logoutPressed?: boolean
   uptime?: number
   autoSignInBrekekePhone?: boolean
+}
+
+export type RingtonePickerType = {
+  [fileName: string]: boolean
 }
 
 export const getLastSignedInId = async (